bots_factory/app/web/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
pydantic
alembic
python-multipart
bcrypt==3.2.0
Jinja2